refactor(occt-wasm): extract meshOps.ts
Continuation of #919/#920/#921. Extracts the Mesh operations section:
mesh, meshEdges, hasTriangulation, meshShape — all WASM heap pointer
reads and Embind delete() lifecycle preserved verbatim.

Adapter: 4121 → 4040 lines (-81). Cumulative reduction from #919: 301 lines.
